+++ CircletExtended.cs
@@ -18,7 +18,7 @@ namespace CircletExtended
     {
         public const string pluginID = "shudnal.CircletExtended";
         public const string pluginName = "Circlet Extended";
-        public const string pluginVersion = "1.1.5";
+        public const string pluginVersion = "1.1.6";
 
         private readonly Harmony harmony = new Harmony(pluginID);